requirements.txt
python-dotenv==1.0.0
pymongo[srv]==4.6.1
openai==1.8.0
langchain==0.1.1
pypdf==4.0.0
gradio==4.15.0
tiktoken==0.5.2